hw1/code/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

art==6.1
gmpy2==2.1.5
numpy==1.26.4
pwn==1.0
sympy==1.12
tqdm==4.66.2
pycryptodome==3.20.0
